HeaderView class added
